# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NJ Upholstery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ServiceCategory = typeof ServiceCategory[keyof typeof ServiceCategory];


export const ServiceCategory = {
  car_seats: 'car_seats',
  couches: 'couches',
  office_chairs: 'office_chairs',
  headboards: 'headboards',
} as const;

export interface Service {
  id: number;
  name: string;
  description: string;
  category: ServiceCategory;
  startingPrice: number;
  imageUrl: string;
  features?: string[];
  /** @nullable */
  turnaroundDays?: number | null;
}

export interface GalleryItem {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  /** @nullable */
  beforeImageUrl?: string | null;
  /** @nullable */
  afterImageUrl?: string | null;
  serviceId: number;
  serviceName: string;
  tags?: string[];
}

export interface Review {
  id: number;
  customerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  /** @nullable */
  serviceId?: number | null;
  /** @nullable */
  serviceName?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  /** @minLength 1 */
  customerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minLength 5 */
  comment: string;
  /** @nullable */
  serviceId?: number | null;
}

export interface ReviewsResponse {
  reviews: Review[];
  averageRating: number;
  totalCount: number;
}

export interface SearchResults {
  query: string;
  services: Service[];
  gallery: GalleryItem[];
}

export type QuoteInputPreferredContact = typeof QuoteInputPreferredContact[keyof typeof QuoteInputPreferredContact];


export const QuoteInputPreferredContact = {
  whatsapp: 'whatsapp',
  call: 'call',
  email: 'email',
} as const;

export interface QuoteInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 7 */
  phone: string;
  /** @nullable */
  email?: string | null;
  serviceCategory: string;
  /** @minLength 10 */
  description: string;
  preferredContact?: QuoteInputPreferredContact;
}

export interface Quote {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  serviceCategory: string;
  description: string;
  /** @nullable */
  preferredContact?: string | null;
  createdAt: string;
}

export interface Stats {
  projectsCompleted: number;
  yearsExperience: number;
  happyClients: number;
  averageRating: number;
}

export type ListGalleryParams = {
/**
 * @nullable
 */
serviceId?: number | null;
};

export type SearchParams = {
q: string;
};