# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NJ Upholstery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List portfolio gallery items
      parameters:
        - name: serviceId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    get:
      operationId: getGalleryItem
      tags: [gallery]
      summary: Get a single gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gallery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
        "404":
          description: Not found
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, description, category, startingPrice, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [car_seats, couches, office_chairs, headboards]
        startingPrice:
          type: number
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
        turnaroundDays:
          type: ["integer", "null"]
    GalleryItem:
      type: object
      required: [id, title, imageUrl, serviceId, serviceName]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: string
        beforeImageUrl:
          type: ["string", "null"]
        afterImageUrl:
          type: ["string", "null"]
        serviceId:
          type: integer
        serviceName:
          type: string
        tags:
          type: array
          items:
            type: string
...