# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NJ Upholstery API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from './serviceCategory';

export interface Service {
  id: number;
  name: string;
  description: string;
  category: ServiceCategory;
  startingPrice: number;
  imageUrl: string;
  features?: string[];
  /** @nullable */
  turnaroundDays?: number | null;
}