# lib/db/src/schema/services.ts
import { pgTable, text, serial, integer, real, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const servicesTable = pgTable("services", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  description: text("description").notNull(),
  category: text("category").notNull(),
  startingPrice: real("starting_price").notNull(),
  imageUrl: text("image_url").notNull(),
  features: jsonb("features").$type<string[]>().notNull().default([]),
  turnaroundDays: integer("turnaround_days"),
});

export const insertServiceSchema = createInsertSchema(servicesTable).omit({ id: true });
export type InsertService = z.infer<typeof insertServiceSchema>;
export type Service = typeof servicesTable.$inferSelect;